requirements.txt
aiogram==3.15.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
httpx==0.28.1
pandas==2.2.3
numpy==2.2.0
pydantic==2.9.2
pydantic-settings==2.6.1
feedparser==6.0.11
APScheduler==3.10.4
python-dateutil==2.9.0.post0
orjson==3.10.12
pytest==8.3.4
pytest-asyncio==0.24.0
respx==0.22.0
openai>=1.0.0
Pillow==11.0.0
